refactor(client): remove entities folder from client and redux folder from src/helpers
