--- packages/client/src/entities/redux/index.js
@@ -1,1 +1,0 @@
-export * from './selectors';
