--- packages/client/src/entities/redux/selectors/index.js
@@ -1,15 +1,0 @@
-/**
- * @module entities/selectors
- * @category Entities
- * @subcategory Selectors
- */
-
-export * from './benefits';
-export * from './brands';
-export * from './categories';
-export * from './entity';
-export * from './facets';
-export * from './merchants';
-export * from './preferences';
-export * from './products';
-export * from './preferences';
