--- packages/client/src/entities/schemas/contact.js
@@ -1,3 +1,0 @@
-import { schema } from 'normalizr';
-
-export default new schema.Entity('contacts');
