--- packages/client/src/entities/schemas/membership.js
@@ -1,3 +1,0 @@
-import { schema } from 'normalizr';
-
-export default new schema.Entity('memberships');
