--- packages/client/src/entities/schemas/merchant.js
@@ -1,4 +1,0 @@
-// @TODO: remove this file
-import { schema } from 'normalizr';
-
-export default new schema.Entity('merchants');
