--- packages/client/src/entities/schemas/statement.js
@@ -1,3 +1,0 @@
-import { schema } from 'normalizr';
-
-export default new schema.Entity('statements');
