--- packages/client/src/entities/schemas/titles.js
@@ -1,5 +1,0 @@
-import { schema } from 'normalizr';
-
-const titles = new schema.Entity('titles');
-
-export default new schema.Object({ entries: [titles] });
