--- packages/client/src/entities/schemas/tracking.js
@@ -1,8 +1,0 @@
-import { schema } from 'normalizr';
-import labelTracking from './labelTracking';
-
-export default new schema.Object(
-  'trackings',
-  { labelTrackings: [labelTracking] },
-  {},
-);
