+++ packages/redux/src/entities/schemas/labelTracking.ts
@@ -1,5 +1,5 @@
 import { schema } from 'normalizr';
-import courier from '@farfetch/blackout-client/entities/schemas/courier';
+import courier from './courier';
 
 export default new schema.Entity(
   'labelTracking',
