+++ packages/redux/src/entities/schemas/order.ts
@@ -1,6 +1,6 @@
 import { adaptDate } from '@farfetch/blackout-client/helpers/adapters';
 import { schema } from 'normalizr';
-import merchant from '@farfetch/blackout-client/entities/schemas/merchant';
+import merchant from './merchant';
 
 export default new schema.Entity(
   'orders',
