+++ packages/redux/src/entities/selectors/entity.ts
@@ -1,3 +1,4 @@
+import get from 'lodash/get';
 import invariant from 'invariant';
 import type { StoreState } from '../../types';
 
@@ -59,3 +60,34 @@ export const getEntityById = <
   id: EntityId,
 ): NonNullable<StateEntities>[SchemaName][EntityId] | undefined =>
   getEntities(state, name)?.[id];
+
+/**
+ * Returns a specific entity by given name and id.
+ *
+ * @function getEntity
+ * @memberof module:entities/selectors
+ *
+ * @param   {object} state - Aplication state.
+ * @param   {string} name  - Entity name.
+ * @param   {string|number} id    - Entity identifier.
+ *
+ * @returns {object} The entity.
+ */
+export const getEntity = (
+  state: StoreState,
+  name: string,
+  id: string | null = null,
+) => {
+  invariant(
+    typeof state.entities !== 'undefined',
+    [
+      'It looks like you are trying to access the `entities` state property but it does not exist.',
+      'You need to add a `entities` reducer to the top level of your state tree.',
+    ].join('\n'),
+  );
+
+  const path = ['entities', name];
+  if (id !== null) path.push(id);
+
+  return get(state, path);
+};
