+++ packages/redux/src/returns/__tests__/selectors.test.ts
@@ -1,4 +1,4 @@
-import * as fromEntities from '@farfetch/blackout-client/entities/redux/selectors/entity';
+import * as fromEntities from '../../entities/selectors/entity';
 import * as fromReturns from '../reducer';
 import * as selectors from '../selectors';
 import {
